A program that finds and prints the first 98 Fibonacci numbers, starting with 1 and 2, followed by a new line

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 /**
- * print_fibonacci_numbers()- prints the first 50 Fibonacci numbers,
+ * print_fibonacci_numbers()- prints the first 98 Fibonacci numbers,
  * starting with 1 and 2
  *
  * Return: Nothing
@@ -10,7 +10,7 @@
 void print_fibonacci_numbers(void)
 {
 	int l;
-	long int a = 0, sum = 1;
+	double a = 0, sum = 1;
 
 	for (l = 0; l < 98; l++)
 	{
@@ -18,10 +18,10 @@ void print_fibonacci_numbers(void)
 		a = sum - a;
 		if (l == 97)
 		{
-			printf("%ld", sum);
+			printf("%.0f", sum);
 		}
 		else
-			printf("%ld, ", sum);
+			printf("%.0f, ", sum);
 
 	}
 	printf("\n");
